Moved Content and Connection in a seperate folder
