+++ src/digitalIdentity-backend/src/main/java/didentity/amos/digitalIdentity/controller/ConnectionController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import didentity.amos.digitalIdentity.services.AuthenticationService;
 import didentity.amos.digitalIdentity.services.DIConnectionService;
-import didentity.amos.digitalIdentity.model.Connection;
+import didentity.amos.digitalIdentity.model.connection.Connection;
 import didentity.amos.digitalIdentity.model.User;
 
 @Controller
