+++ src/digitalIdentity-backend/src/main/java/didentity/amos/digitalIdentity/messages/responses/ConnectionsResponse.java
@@ -11,7 +11,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-import didentity.amos.digitalIdentity.model.Content;
+import didentity.amos.digitalIdentity.model.connection.Content;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
