+++ src/digitalIdentity-backend/src/main/java/didentity/amos/digitalIdentity/model/connection/Connection.java
@@ -1,4 +1,4 @@
-package didentity.amos.digitalIdentity.model;
+package didentity.amos.digitalIdentity.model.connection;
 
 import didentity.amos.digitalIdentity.enums.UserRole;
