+++ src/digitalIdentity-backend/src/main/java/didentity/amos/digitalIdentity/services/DIConnectionService.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClientException;
 
 import didentity.amos.digitalIdentity.enums.UserRole;
-import didentity.amos.digitalIdentity.model.Connection;
-import didentity.amos.digitalIdentity.model.Content;
+import didentity.amos.digitalIdentity.model.connection.Connection;
+import didentity.amos.digitalIdentity.model.connection.Content;
 import didentity.amos.digitalIdentity.messages.responses.ConnectionsResponse;
 import didentity.amos.digitalIdentity.messages.responses.CreateConnectionResponse;
 import didentity.amos.digitalIdentity.model.User;
